Refined TypeScript interfaces and shared type 'UserType' as follows:
Changes in 'Rating' Interface:

Removed unnecessary fields, streamlining the interface to meet the updated requirements.
Renamed 'customerId' to 'raterUserId' for consistency.
Renamed 'businessId' to 'ratedUserId' to align with new naming conventions.
Updated 'bikerId' to 'raterUserId' for uniformity.
Adjusted comments to reflect the field name changes.
Changes in 'AverageRating' Interface:

Removed unnecessary fields, simplifying the interface in accordance with the revised specifications.
Renamed 'businessId' to 'ratedUserId' to match the updated requirements.
Updated comments to document the modifications accurately.
Addition to 'shared' Module:

Introduced the export type 'UserType' with possible values 'biker,' 'customer,' and 'business' for shared use across the codebase.
These adjustments improve code efficiency by eliminating unnecessary fields and ensuring that the interfaces align with the updated requirements. The addition of 'UserType' enhances code readability and maintainability.
