+++ docs/interfaces/AdminUser.md
@@ -28,7 +28,7 @@ rol del admin - ver [AdminRole](../modules.md#adminrole)
 
 #### Defined in
 
-[src/admin/admin.ts:28](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L28)
+[src/admin/admin.ts:28](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L28)
 
 ___
 
@@ -40,7 +40,7 @@ fecha y hora de creacion
 
 #### Defined in
 
-[src/admin/admin.ts:31](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L31)
+[src/admin/admin.ts:31](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L31)
 
 ___
 
@@ -52,7 +52,7 @@ correo electronico
 
 #### Defined in
 
-[src/admin/admin.ts:22](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L22)
+[src/admin/admin.ts:22](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L22)
 
 ___
 
@@ -64,7 +64,7 @@ id de firebase auth
 
 #### Defined in
 
-[src/admin/admin.ts:13](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L13)
+[src/admin/admin.ts:13](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L13)
 
 ___
 
@@ -76,7 +76,7 @@ nombre
 
 #### Defined in
 
-[src/admin/admin.ts:16](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L16)
+[src/admin/admin.ts:16](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L16)
 
 ___
 
@@ -88,7 +88,7 @@ apellidos
 
 #### Defined in
 
-[src/admin/admin.ts:19](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L19)
+[src/admin/admin.ts:19](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L19)
 
 ___
 
@@ -100,7 +100,7 @@ fecha y hora de ultima actualizacion
 
 #### Defined in
 
-[src/admin/admin.ts:34](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L34)
+[src/admin/admin.ts:34](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L34)
 
 ___
 
@@ -112,7 +112,7 @@ URL de la foto de perfil (almacenada en Firebase Storage)
 
 #### Defined in
 
-[src/admin/admin.ts:25](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L25)
+[src/admin/admin.ts:25](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L25)
 
 ___
 
@@ -124,4 +124,4 @@ estado de la cuenta de usario administrado (si fue o no aprobado por un usuario 
 
 #### Defined in
 
-[src/admin/admin.ts:37](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L37)
+[src/admin/admin.ts:37](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L37)
