+++ docs/interfaces/Coords.md
@@ -19,7 +19,7 @@ Coordenadas
 
 #### Defined in
 
-[src/other/shared.ts:12](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/other/shared.ts#L12)
+[src/other/shared.ts:12](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/other/shared.ts#L12)
 
 ___
 
@@ -29,4 +29,4 @@ ___
 
 #### Defined in
 
-[src/other/shared.ts:11](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/other/shared.ts#L11)
+[src/other/shared.ts:11](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/other/shared.ts#L11)
