+++ docs/interfaces/Customer.md
@@ -29,7 +29,7 @@ numero de celular del usuario
 
 #### Defined in
 
-[src/customer/customer.ts:23](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L23)
+[src/customer/customer.ts:23](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L23)
 
 ___
 
@@ -41,7 +41,7 @@ fecha y hora de creacion
 
 #### Defined in
 
-[src/customer/customer.ts:32](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L32)
+[src/customer/customer.ts:32](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L32)
 
 ___
 
@@ -53,7 +53,7 @@ ubicacion actual del negocio, utilizado para calcular rutas de los envios
 
 #### Defined in
 
-[src/customer/customer.ts:29](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L29)
+[src/customer/customer.ts:29](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L29)
 
 ___
 
@@ -65,7 +65,7 @@ correo electronico del usuario
 
 #### Defined in
 
-[src/customer/customer.ts:14](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L14)
+[src/customer/customer.ts:14](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L14)
 
 ___
 
@@ -77,7 +77,7 @@ token para los mensajes de la nube
 
 #### Defined in
 
-[src/customer/customer.ts:20](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L20)
+[src/customer/customer.ts:20](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L20)
 
 ___
 
@@ -89,7 +89,7 @@ id de firebase auth
 
 #### Defined in
 
-[src/customer/customer.ts:17](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L17)
+[src/customer/customer.ts:17](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L17)
 
 ___
 
@@ -101,7 +101,7 @@ nombre completo del usuario
 
 #### Defined in
 
-[src/customer/customer.ts:8](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L8)
+[src/customer/customer.ts:8](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L8)
 
 ___
 
@@ -113,7 +113,7 @@ apellidos del usuario
 
 #### Defined in
 
-[src/customer/customer.ts:11](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L11)
+[src/customer/customer.ts:11](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L11)
 
 ___
 
@@ -125,7 +125,7 @@ fecha y hora de ultima actualizacion
 
 #### Defined in
 
-[src/customer/customer.ts:35](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L35)
+[src/customer/customer.ts:35](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L35)
 
 ___
 
@@ -137,4 +137,4 @@ URL de la foto de perfil (almacenada en Firebase Storage)
 
 #### Defined in
 
-[src/customer/customer.ts:26](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/customer/customer.ts#L26)
+[src/customer/customer.ts:26](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/customer/customer.ts#L26)
