+++ docs/modules.md
@@ -33,6 +33,7 @@
 - [OrderStatus](modules.md#orderstatus)
 - [PaymentMethod](modules.md#paymentmethod)
 - [UserStatus](modules.md#userstatus)
+- [UserType](modules.md#usertype)
 
 ### Variables
 
@@ -49,7 +50,7 @@ Tipo de usuario adminsitrador
 
 #### Defined in
 
-[src/admin/admin.ts:6](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/admin/admin.ts#L6)
+[src/admin/admin.ts:6](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/admin/admin.ts#L6)
 
 ___
 
@@ -61,7 +62,7 @@ Si el usuario ciclista esta disponible o no para hacer entregas
 
 #### Defined in
 
-[src/biker/biker.ts:4](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/biker/biker.ts#L4)
+[src/biker/biker.ts:4](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/biker/biker.ts#L4)
 
 ___
 
@@ -73,7 +74,7 @@ tipo del contacto adicional, como redes sociales o sitio web
 
 #### Defined in
 
-[src/business/business.ts:5](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/business/business.ts#L5)
+[src/business/business.ts:5](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/business/business.ts#L5)
 
 ___
 
@@ -96,7 +97,7 @@ Canceled: el pedido es cancelado por el administrador
 
 #### Defined in
 
-[src/order/order.ts:17](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/order/order.ts#L17)
+[src/order/order.ts:18](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/order/order.ts#L18)
 
 ___
 
@@ -108,7 +109,7 @@ Metodos de pago
 
 #### Defined in
 
-[src/order/order.ts:46](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/order/order.ts#L46)
+[src/order/order.ts:47](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/order/order.ts#L47)
 
 ___
 
@@ -121,7 +122,17 @@ Bikers, Admin y Business
 
 #### Defined in
 
-[src/other/shared.ts:5](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/other/shared.ts#L5)
+[src/other/shared.ts:5](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/other/shared.ts#L5)
+
+___
+
+### UserType
+
+Ƭ **UserType**: ``"biker"`` \| ``"customer"`` \| ``"business"``
+
+#### Defined in
+
+[src/other/shared.ts:15](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/other/shared.ts#L15)
 
 ## Variables
 
@@ -133,7 +144,7 @@ Métodos de pago en español
 
 #### Defined in
 
-[src/order/order.ts:49](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/order/order.ts#L49)
+[src/order/order.ts:50](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/order/order.ts#L50)
 
 ___
 
@@ -145,4 +156,4 @@ Status de la Orden en español
 
 #### Defined in
 
-[src/order/order.ts:31](https://github.com/gatitolabs/ride-dbtypes/blob/5d6cff1/src/order/order.ts#L31)
+[src/order/order.ts:32](https://github.com/gatitolabs/ride-dbtypes/blob/1de9b75/src/order/order.ts#L32)
