+++ lib/index.d.ts
@@ -3,12 +3,13 @@ export { Biker, BikerStatus } from './src/biker/biker';
 export { Business, BusinessContact, BusinessContactType } from './src/business/business';
 export { Customer } from './src/customer/customer';
 export { Order, OrderItem, OrderStatus, PaymentMethod, statusInSpanish, paymentMethodInSpanish } from './src/order/order';
-export { Rating, AverageRating } from './src/order/rating';
+export { Rating } from './src/rating/rating';
+export { AverageRating } from './src/rating/averageRating';
 export { DeliveryAvailable } from './src/other/deliveryAvailable';
 export { Product } from './src/product/product';
 export { Tag } from './src/product/tag';
 export { ConfigItem } from './src/admin/configItem';
-export { UserStatus, Coords } from './src/other/shared';
+export { UserStatus, Coords, UserType } from './src/other/shared';
 /**
  * Colecciones de Firestore
  */
