+++ lib/src/rating/rating.js
